requirements.txt
# Install PyTorch / torchvision separately first when needed.
# Example (desktop CUDA):
#   pip install torch torchvision --index-url https://download.pytorch.org/whl/cu128
# Example (CPU):
#   pip install torch torchvision
#
# On Jetson, prefer NVIDIA/Jetson-compatible PyTorch wheels instead of plain pip wheels.

numpy==2.4.3
opencv-python==4.13.0.0
pandas==3.0.2
Pillow==12.1.1
scipy==1.17.1
tqdm==4.67.3
PyYAML==6.0.3
matplotlib==3.10.7
einops==0.8.1
h5py==3.15.1
seaborn==0.13.2
packaging==25.0
ipython==9.7.0
